+++ app/src/test/java/pm/gnosis/heimdall/data/repositories/impls/WalletConnectBridgeRepositoryTest.kt
@@ -1,17 +1,22 @@
 package pm.gnosis.heimdall.data.repositories.impls
 
 import com.squareup.moshi.Moshi
+import com.squareup.moshi.Types
 import io.reactivex.rxkotlin.subscribeBy
 import okhttp3.OkHttpClient
 import org.junit.Rule
 import org.junit.Test
 import pm.gnosis.heimdall.BuildConfig
+import pm.gnosis.heimdall.data.repositories.impls.wc.WCSessionStore
 import pm.gnosis.heimdall.data.repositories.models.SafeTransaction
 import pm.gnosis.tests.utils.ImmediateSchedulersRule
 import pm.gnosis.utils.asEthereumAddress
+import pm.gnosis.utils.nullOnThrow
+import java.io.File
 import java.lang.IllegalStateException
 import java.net.URL
 import java.util.*
+import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.TimeUnit
 
 class WalletConnectBridgeRepositoryTest {
@@ -22,7 +27,9 @@ class WalletConnectBridgeRepositoryTest {
 
     @Test
     fun integration() {
-        val repo = WalletConnectBridgeRepository(OkHttpClient.Builder().pingInterval(1000, TimeUnit.MILLISECONDS).build(), Moshi.Builder().build())
+        val client = OkHttpClient.Builder().pingInterval(1000, TimeUnit.MILLISECONDS).build()
+        val moshi = Moshi.Builder().build()
+        val repo = WalletConnectBridgeRepository(client, moshi, FileWCSessionStore(moshi))
         val uri =
             "wc:645bf32b-8697-476f-a85f-7125a03bd012@1?bridge=https%3A%2F%2Fbridge.walletconnect.org&key=98650b0362374ef67e5d434900fdf294d861480ee180ea50ea080476c926ef5a"
 
@@ -91,4 +98,43 @@ class WalletConnectBridgeRepositoryTest {
             }
         )
     }
+}
+
+class FileWCSessionStore(moshi: Moshi) : WCSessionStore {
+    private val adapter = moshi.adapter<Map<String, WCSessionStore.State>>(
+        Types.newParameterizedType(
+            Map::class.java,
+            String::class.java,
+            WCSessionStore.State::class.java
+        )
+    )
+
+    private val storageFile: File = File("build/tmp/test_store.json").apply { createNewFile() }
+    private val currentStates: MutableMap<String, WCSessionStore.State> = ConcurrentHashMap()
+
+    init {
+        val storeContent = storageFile.readText()
+        nullOnThrow { adapter.fromJson(storeContent) }?.let {
+            currentStates.putAll(it)
+        }
+    }
+
+    override fun load(id: String): WCSessionStore.State? = currentStates[id]
+
+    override fun store(id: String, state: WCSessionStore.State) {
+        currentStates[id] = state
+        writeToFile()
+    }
+
+    override fun remove(id: String) {
+        currentStates.remove(id)
+        writeToFile()
+    }
+
+    override fun list(): List<WCSessionStore.State> = currentStates.values.toList()
+
+    private fun writeToFile() {
+        storageFile.writeText(adapter.toJson(currentStates))
+    }
+
 }
